Refactor channelService
