+++ v1/services/channelService.go
@@ -1,12 +1,14 @@
-package service
+package services
 
 import (
 	"Osheet-api/database"
 	"Osheet-api/v1/models"
 )
 
+type ChannelService struct{}
+
 // Get all channels
-func GetAllChannel() (channelList []*models.Channel, err error) {
+func (s ChannelService) GetAllChannel() (channelList []*models.Channel, err error) {
 	if err = database.GetDB().Find(&channelList).Error; err != nil {
 		return nil, err
 	}
@@ -15,7 +17,7 @@ func GetAllChannel() (channelList []*models.Channel, err error) {
 }
 
 // Get channel by twiiter account
-func GetChannelByTwitterAccount(twitterAccount string) (channel *models.Channel, err error) {
+func (s ChannelService) GetChannelByTwitterAccount(twitterAccount string) (channel *models.Channel, err error) {
 	if err = database.GetDB().Where("twitter_account = ?", twitterAccount).First(&channel).Error; err != nil {
 		return nil, err
 	}
